removed defaults module and simplified hook object
