+++ src/tmtccmd/config/logo.png
