+++ src/tmtccmd/pus_tm/factory.py
@@ -29,3 +29,14 @@ class PusTelemetryFactory(object):
             LOGGER.error("PusTelemetryFactory:create: Invalid packet format")
             return None
 
+
+def default_factory_hook(raw_tm_packet: bytearray) -> PusTelemetry:
+    service_type = raw_tm_packet[7]
+    if service_type == 1:
+        return Service1TM(raw_tm_packet)
+    if service_type == 5:
+        return Service5TM(raw_tm_packet)
+    if service_type == 17:
+        return Service17TM(raw_tm_packet)
+    LOGGER.info("The service " + str(service_type) + " is not implemented in Telemetry Factory")
+    return PusTelemetry(raw_tm_packet)
